Moving code to lightningbase
